add Rand trait
