+++ field/src/extension_field/quadratic.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::extension_field::{Extendable, FieldExtension, Frobenius, OEF};
 use crate::field_types::Field;
-use crate::ops::Square;
+use crate::ops::{Square, Rand};
 
 #[derive(Copy, Clone, Eq, PartialEq, Hash, Serialize, Deserialize)]
 #[serde(bound = "")]
@@ -102,7 +102,9 @@ impl<F: Extendable<2>> Field for QuadraticExtension<F> {
     fn from_noncanonical_u128(n: u128) -> Self {
         F::from_noncanonical_u128(n).into()
     }
+}
 
+impl<F: Extendable<2>> Rand for QuadraticExtension<F> {
     fn rand_from_rng<R: Rng>(rng: &mut R) -> Self {
         Self([F::rand_from_rng(rng), F::rand_from_rng(rng)])
     }
