+++ field/src/extension_field/quintic.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::extension_field::{Extendable, FieldExtension, Frobenius, OEF};
 use crate::field_types::Field;
-use crate::ops::Square;
+use crate::ops::{Rand, Square};
 
 #[derive(Copy, Clone, Eq, PartialEq, Hash, Serialize, Deserialize)]
 #[serde(bound = "")]
@@ -111,7 +111,9 @@ impl<F: Extendable<5>> Field for QuinticExtension<F> {
     fn from_noncanonical_u128(n: u128) -> Self {
         F::from_noncanonical_u128(n).into()
     }
+}
 
+impl<F: Extendable<5>> Rand for QuinticExtension<F> {
     fn rand_from_rng<R: Rng>(rng: &mut R) -> Self {
         Self::from_basefield_array([
             F::rand_from_rng(rng),
