+++ plonky2/benches/merkle.rs
@@ -2,9 +2,9 @@
 
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use plonky2::field::goldilocks_field::GoldilocksField;
+use plonky2::hash::blake3::Blake3Hash;
 use plonky2::hash::hash_types::RichField;
 use plonky2::hash::keccak::KeccakHash;
-use plonky2::hash::blake3::Blake3Hash;
 use plonky2::hash::merkle_tree::MerkleTree;
 use plonky2::hash::poseidon::PoseidonHash;
 use plonky2::plonk::config::Hasher;
