+++ plonky2/benches/reverse_index_bits.rs
@@ -1,6 +1,7 @@
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use plonky2::field::field_types::Field;
 use plonky2::field::goldilocks_field::GoldilocksField;
+use plonky2_field::ops::Rand;
 use plonky2_util::{reverse_index_bits, reverse_index_bits_in_place};
 
 type F = GoldilocksField;
