+++ plonky2/benches/transpose.rs
@@ -1,6 +1,7 @@
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use plonky2::field::field_types::Field;
 use plonky2::field::goldilocks_field::GoldilocksField;
+use plonky2::field::ops::Rand;
 use plonky2::util::transpose;
 
 fn criterion_benchmark(c: &mut Criterion) {
