+++ plonky2/src/gadgets/arithmetic_extension.rs
@@ -561,6 +561,7 @@ mod tests {
     use anyhow::Result;
     use plonky2_field::extension_field::algebra::ExtensionAlgebra;
     use plonky2_field::field_types::Field;
+    use plonky2_field::ops::Rand;
 
     use crate::iop::ext_target::ExtensionAlgebraTarget;
     use crate::iop::witness::{PartialWitness, Witness};
