+++ plonky2/src/gadgets/random_access.rs
@@ -59,6 +59,7 @@ impl<F: RichField + Extendable<D>, const D: usize> CircuitBuilder<F, D> {
 mod tests {
     use anyhow::Result;
     use plonky2_field::field_types::Field;
+    use plonky2_field::ops::Rand;
 
     use super::*;
     use crate::iop::witness::PartialWitness;
