+++ plonky2/src/gates/gate_testing.rs
@@ -1,6 +1,7 @@
 use anyhow::{ensure, Result};
 use plonky2_field::extension_field::{Extendable, FieldExtension};
 use plonky2_field::field_types::Field;
+use plonky2_field::ops::Rand;
 use plonky2_field::polynomial::{PolynomialCoeffs, PolynomialValues};
 use plonky2_util::log2_ceil;
