+++ plonky2/src/gates/interpolation.rs
@@ -277,6 +277,7 @@ mod tests {
     use anyhow::Result;
     use plonky2_field::field_types::Field;
     use plonky2_field::goldilocks_field::GoldilocksField;
+    use plonky2_field::ops::Rand;
     use plonky2_field::polynomial::PolynomialCoeffs;
 
     use crate::gadgets::interpolation::InterpolationGate;
