+++ plonky2/src/gates/random_access.rs
@@ -367,6 +367,7 @@ mod tests {
     use anyhow::Result;
     use plonky2_field::field_types::Field;
     use plonky2_field::goldilocks_field::GoldilocksField;
+    use plonky2_field::ops::Rand;
     use rand::{thread_rng, Rng};
 
     use crate::gates::gate::Gate;
