+++ plonky2/src/hash/hash_types.rs
@@ -1,5 +1,6 @@
 use plonky2_field::field_types::{Field, PrimeField64};
 use plonky2_field::goldilocks_field::GoldilocksField;
+use plonky2_field::ops::Rand;
 use rand::Rng;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
@@ -47,10 +48,17 @@ impl<F: Field> HashOut<F> {
             ],
         }
     }
+}
 
-    pub fn rand() -> Self {
+impl<F: Field> Rand for HashOut<F> {
+    fn rand_from_rng<R: Rng>(rng: &mut R) -> Self {
         Self {
-            elements: [F::rand(), F::rand(), F::rand(), F::rand()],
+            elements: [
+                F::rand_from_rng(rng),
+                F::rand_from_rng(rng),
+                F::rand_from_rng(rng),
+                F::rand_from_rng(rng),
+            ],
         }
     }
 }
@@ -114,15 +122,11 @@ pub struct MerkleCapTarget(pub Vec<HashOutTarget>);
 #[derive(Eq, PartialEq, Copy, Clone, Debug)]
 pub struct BytesHash<const N: usize>(pub [u8; N]);
 
-impl<const N: usize> BytesHash<N> {
-    pub fn rand_from_rng<R: Rng>(rng: &mut R) -> Self {
+impl<const N: usize> Rand for BytesHash<N> {
+    fn rand_from_rng<R: Rng>(rng: &mut R) -> Self {
         let mut buf = [0; N];
         rng.fill_bytes(&mut buf);
         Self(buf)
-    }
-
-    pub fn rand() -> Self {
-        Self::rand_from_rng(&mut rand::thread_rng())
     }
 }
