+++ plonky2/src/hash/path_compression.rs
@@ -116,6 +116,7 @@ where
 #[cfg(test)]
 mod tests {
     use plonky2_field::field_types::Field;
+    use plonky2_field::ops::Rand;
     use rand::{thread_rng, Rng};
 
     use super::*;
