+++ plonky2/src/iop/challenger.rs
@@ -275,6 +275,7 @@ impl<F: RichField + Extendable<D>, H: AlgebraicHasher<F>, const D: usize>
 #[cfg(test)]
 mod tests {
     use plonky2_field::field_types::Field;
+    use plonky2_field::ops::Rand;
 
     use crate::iop::challenger::{Challenger, RecursiveChallenger};
     use crate::iop::generator::generate_partial_witness;
